下单支持碎股交易，兼容原来下单方法

+++ src/main/java/com/tigerbrokers/stock/openapi/client/util/builder/TradeParamBuilder.java
@@ -107,7 +107,7 @@ public class TradeParamBuilder {
     return this;
   }
 
-  public TradeParamBuilder totalQuantity(Integer totalQuantity) {
+  public TradeParamBuilder totalQuantity(Long totalQuantity) {
     if (totalQuantity != null) {
       this.orderParameter.setTotalQuantity(totalQuantity);
     }
